% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpressionRecovery.R
\name{GeneExpressionRecovery}
\alias{GeneExpressionRecovery}
\title{Gene Expression Recovery}
\usage{
GeneExpressionRecovery(InputDir, Donors, FC)
}
\arguments{
\item{InputDir}{#'A folder including gene-cell count matrix (read_count.csv), filtered feature-barcode matrices from 10X CellRanger count (barcodes.tsv.gz), list of QC passed barcodes (Cells_passed_QC_noDoublet.txt) in QC/ folder, and output of genetic demultiplexing of the sample pool using vireo (donor_ids.tsv).}

\item{Donors}{A pair of selected donors (e.g., donor6_donor2)}

\item{FC}{A name for flow cell (e.g., FAI5649A17).
#'@return You can find the results in the SAVER/ folder with 'AssignedCells.txt' and 'AllCells.txt' extensions.}
}
\description{
This function uses SAVER (single-cell analysis via expression recovery), an expression recovery method for unique molecule index (UMI)-based scRNA-seq data to provide accurate expression estimates for all genes in as scRNA-seq profile.
}
\examples{
library("scDIV")
csQCEAdir <- system.file("extdata", package = "scDIV")
Donors='donor6_donor2'
FC='FAI5649A17'
GeneExpressionRecovery( InputDir = csQCEAdir, Donors = Donors, FC = FC )
}
\author{
{Isar Nassiri}
}
