% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDCA.R
\name{IDCA}
\alias{IDCA}
\title{Inter-individual Differential gene Correlation Analysis (IDCA)}
\usage{
IDCA(InputDir, Donors, FC, ERP, TEST = FALSE)
}
\arguments{
\item{InputDir}{A folder including gene-cell count matrix (read_count.csv), filtered feature-barcode matrices from 10X CellRanger count (barcodes.tsv.gz), and output of genetic demultiplexing of the sample pool using vireo (donor_ids.tsv).}

\item{Donors}{A pair of selected donors (e.g., donor6_donor2)}

\item{FC}{A name for flow cell (e.g., FAI5649A17).}

\item{ERP}{Matrix of cell-gene expression after Gene Expression Recovery in the SAVER/ folder with 'AssignedCells.txt' and 'AllCells.txt' extensions.}

\item{TEST}{use TEST = TRUE to get sample outputs.}
}
\value{
You can find the results in the IDCA_Analysis/ folder.
}
\description{
The functions uses correlation coefficients and performed Inter-individual Differential gene Correlation Analysis (IDCA) for two donors (D1 and D2) and genes (G1 and G2).
}
\examples{
library("scDIV")
ERP = "donor6_donor2_FAI5649A17_AssignedCells.txt"
Donors='donor6_donor2'
FC='FAI5649A17'
InputDir = system.file("extdata", package = "scDIV")
IDCA( InputDir, Donors, FC, ERP, TEST = TRUE )
}
\author{
{Isar Nassiri}
}
