% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDCAvis.R
\name{IDCAvis}
\alias{IDCAvis}
\title{Visualization of IDCA outputs of Inter-individual Differential gene Correlation Analysis (IDCA) results}
\usage{
IDCAvis(InputDir)
}
\arguments{
\item{InputDir}{A folder including matrix of cell-gene expression after Gene Expression Recovery in the SAVER/ folder with 'AssignedCells.txt' extension, output of genetic demultiplexing of the sample pool using vireo (donor_ids.tsv), and output of IDCA function with extension of "_IDCA.txt" in IDCA_Analysis/ folder.}
}
\value{
You can find the results in the IDCA_Analysis/IDCA_Plots/ as pdf file(s).
}
\description{
The function visualizes the outputs of IDC analysis.
}
\examples{
library("scDIV")
InputDir = system.file("extdata", package = "scDIV")
IDCAvis( InputDir )
}
\author{
{Isar Nassiri}
}
