% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scQC.R
\name{scQC}
\alias{scQC}
\title{Single Cell data Quality Control}
\usage{
scQC(InputDir)
}
\arguments{
\item{InputDir}{#'A folder including gene-cell count matrix (read_count.csv) and filtered feature-barcode matrices from 10X CellRanger count (barcodes.tsv.gz)}
}
\value{
You can find the results in the QC/ folder under the title of 'Cells_passed_QC_noDoublet.txt'.
}
\description{
Identify potential doublet cells based on the local density of simulated doublet expression profiles using scDblFinder package.
}
\examples{
library("scDIV")
csQCEAdir <- system.file("extdata", package = "scDIV")
scQC( InputDir = csQCEAdir ) 
}
\author{
{Isar Nassiri}
}
